==> notebooks/03 - Interpreting Our Models.ipynb <==
import eli5
import pickle
import pandas as pd
%pylab inline

test_df = pd.read_csv('../data/house_test_final.csv')
test_df = test_df.set_index('Id')
rfr_pred = pd.read_csv('../data/tuned_rfr_output_submission.csv')
svr_pred = pd.read_csv('../data/svr_output_submission.csv')

rfr_pred.describe()

rfr_pred['SalePrice'].hist()

svr_pred.describe()

svr_pred['SalePrice'].hist()

### Discussion

- What explanations might we want to start with? Why?

## Exercise

Pick out the highest value row of the `svr` model and save it as `example`.

%load ../solutions/svr_top.py


example

example_row = test_df.loc[example.Id]

example_row

### Now we need to revive our models

svr = pickle.load(open('../data/models/sgb.sav', 'rb'))

eli5.explain_prediction?

eli5.explain_prediction(svr, example_row)

eli5.explain_weights_df(svr)

len(test_df.columns)

### Discussion

- What do these features mean?
- What do we know about support vector machines?

### Exercise

Build a dictionary with the keys being 'x{col_number}' and the values being the feature name. Call it `feature_names`.

%load ../solutions/feature_name_dict.py


feature_names

prediction_df = eli5.explain_prediction_df(svr, example_row)

prediction_df['feature_name'] = prediction_df.feature.map(
    lambda x: feature_names[x] if x in feature_names else x)

prediction_df.sort_values?

prediction_df.sort_values('weight', ascending=False)

### Discussion

- What are the most impactful values?
- Do these seem valuable? Overvalued? Undervalued?

## Bonus exercise

- Take a look at the overall weights and add the feature name notations. 
- Are these weights different from the particular explanation?

### Investigating another model

- Use the same strategy to evaluate another model. What differences did you find? Which makes more sense to you? Why?

%load ../solutions/eli5_weights.py
